chore(type-designer): reorganize files (add view, drop barrel pattern, rename close to domain)
