+++ packages/plugins/type-designer/src/components/content-card.svelte
