--- packages/plugins/type-designer/src/components/elements-type-container/index.ts
@@ -1,1 +1,0 @@
-export { default as ElementsTypeContainer } from './columns.svelte'
