--- packages/plugins/type-designer/src/components/index.ts
@@ -1,1 +1,0 @@
-export * from './elements-type-container'
