--- packages/plugins/type-designer/src/stores/index.ts
@@ -1,6 +1,0 @@
-// PLUGIN BASE
-export * from './plugin.store'
-// CUSTOM
-
-export * from './data-types-templates.store'
-export * from './columns.store'
