+++ packages/plugins/type-designer/src/stores/types.store.d.ts
